Update kamal secrets file (#1289)
Kamal 2 replaces `.evv.erb` with a `.kamal/secrets`.
In the former you could use erb tags `<%= %>` to
run ruby code, however this is not possible anymore.

This erb tag gets treated as a string.

This took me a while to spot because in its output
sensitive information is redacted for security.