Simplify time zones (#1292)
* Simplify time zones

Initially, we collected time zone on registration. The idea was it would
be helpful to know so you could display times and zones permanently at
the right time for the user.

Since browser time zones are slightly different than Rails, we had to
store it in a cookie, select it on registration, and it felt like a lot
for a seemingly simple feature.

This hasn't proven to be very useful in our applications because
local_time does everything we need. As a result, we figured it's time to
yank this feature. It simplifies registration with one less field and
the users can always set their time zone in their profile if it's still
needed and local_time doesn't work for your application.

* Remove time_zone param